+++ mighty_logger/__init__.py
@@ -16,9 +16,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from .powerful_logger import MightyLogger
+from .mighty_logger import MightyLogger
 from .simple_logger import Logger
 
 __authot__ = "Kalynovsky 'Nakamura Akira' Valentin"
-__version__ = "0.7.1"
+__version__ = "0.7.2"
 __email__ = "nakama3942@gmail.com"